refactor(ui): establish shadcn hygiene baseline
