+++ apps/admin/app/tasks/use-task-form.ts
@@ -1,4 +1,4 @@
-import { useAsymForm } from "@asym/ui/components/shadcn/form";
+import { useAsymForm } from "@asym/ui/components/shadcn/tanstack-form";
 
 import {
   createInitialTaskFormValues,
