+++ apps/admin/src/cms-ui/web-studio/collections/shared/document-workspace/NativeDocumentWorkspaceSettingsDialog.tsx
@@ -9,7 +9,7 @@ import {
   DialogHeader,
   DialogTitle,
 } from "@asym/ui/components/shadcn/dialog";
-import { useAsymForm } from "@asym/ui/components/shadcn/form";
+import { useAsymForm } from "@asym/ui/components/shadcn/tanstack-form";
 import { usePreferences } from "@payloadcms/ui";
 import { useEffect, useState } from "react";
 import { z } from "zod";
