+++ packages/missionary/components/add-partner-dialog.tsx
@@ -9,7 +9,7 @@ import {
   DialogTitle,
   DialogTrigger,
 } from "@asym/ui/components/shadcn/dialog";
-import { useAsymForm } from "@asym/ui/components/shadcn/form";
+import { useAsymForm } from "@asym/ui/components/shadcn/tanstack-form";
 import { Loader2 } from "lucide-react";
 import * as React from "react";
 import { toast } from "sonner";
